+++ payment/datastore.go
@@ -1,0 +1,148 @@
+package payment
+
+import (
+	"os"
+
+	migrate "github.com/golang-migrate/migrate/v4"
+	"github.com/golang-migrate/migrate/v4/database/postgres"
+	"github.com/jmoiron/sqlx"
+	uuid "github.com/satori/go.uuid"
+	"github.com/shopspring/decimal"
+
+	// needed for magic migration
+	_ "github.com/golang-migrate/migrate/v4/source/file"
+)
+
+// Datastore abstracts over the underlying datastore
+type Datastore interface {
+	// CreateOrder is used to create an order for payments
+	CreateOrder(totalPrice decimal.Decimal, merchantID string, status string, orderItems []OrderItem) (*Order, error)
+	// GetOrder by ID
+	GetOrder(orderID uuid.UUID) (*Order, error)
+}
+
+// Postgres is a Datastore wrapper around a postgres database
+type Postgres struct {
+	*sqlx.DB
+}
+
+// NewMigrate creates a Migrate instance given a Postgres instance with an active database connection
+func (pg *Postgres) NewMigrate() (*migrate.Migrate, error) {
+	driver, err := postgres.WithInstance(pg.DB.DB, &postgres.Config{})
+	if err != nil {
+		return nil, err
+	}
+
+	dbMigrationsURL := os.Getenv("DATABASE_MIGRATIONS_URL")
+	m, err := migrate.NewWithDatabaseInstance(
+		dbMigrationsURL,
+		"postgres",
+		driver,
+	)
+	if err != nil {
+		return nil, err
+	}
+
+	return m, err
+}
+
+// Migrate the Postgres instance
+func (pg *Postgres) Migrate() error {
+	m, err := pg.NewMigrate()
+	if err != nil {
+		return err
+	}
+
+	err = m.Migrate(4)
+	if err != migrate.ErrNoChange && err != nil {
+		return err
+	}
+	return nil
+}
+
+// NewPostgres creates a new Postgres Datastore
+func NewPostgres(databaseURL string, performMigration bool) (*Postgres, error) {
+	if len(databaseURL) == 0 {
+		databaseURL = os.Getenv("DATABASE_URL")
+	}
+
+	db, err := sqlx.Open("postgres", databaseURL)
+	if err != nil {
+		return nil, err
+	}
+
+	pg := &Postgres{db}
+
+	if performMigration {
+		err = pg.Migrate()
+		if err != nil {
+			return nil, err
+		}
+	}
+
+	return pg, nil
+}
+
+// CreateOrder given the promotion type, initial number of grants and the desired value of those grants
+func (pg *Postgres) CreateOrder(totalPrice decimal.Decimal, merchantID string, status string, orderItems []OrderItem) (*Order, error) {
+	tx := pg.DB.MustBegin()
+
+	var order Order
+	err := tx.Get(&order, `
+			INSERT INTO orders (total_price, merchant_id, status)
+			VALUES ($1, $2, $3)
+			RETURNING *
+		`,
+		totalPrice, merchantID, status)
+
+	if err != nil {
+		_ = tx.Rollback()
+		return nil, err
+	}
+
+	for i := 0; i < len(orderItems); i++ {
+		orderItems[i].OrderID = order.ID
+
+		nstmt, _ := tx.PrepareNamed(`
+			INSERT INTO order_items (order_id, quantity, price, currency, subtotal)
+			VALUES (:order_id, :quantity, :price, :currency, :subtotal)
+			RETURNING *
+		`)
+		err = nstmt.Get(&orderItems[i], orderItems[i])
+
+		if err != nil {
+			_ = tx.Rollback()
+			return nil, err
+		}
+	}
+	err = tx.Commit()
+	if err != nil {
+		_ = tx.Rollback()
+		return nil, err
+	}
+
+	order.Items = orderItems
+
+	return &order, nil
+}
+
+// GetOrder queries the database and returns an order
+func (pg *Postgres) GetOrder(orderID uuid.UUID) (*Order, error) {
+	statement := "select * from orders where id = $1"
+	order := Order{}
+	err := pg.DB.Get(&order, statement, orderID)
+	if err != nil {
+		return nil, err
+	}
+
+	foundOrderItems := []OrderItem{}
+	statement = "select * from order_items where order_id = $1"
+	err = pg.DB.Select(&foundOrderItems, statement, orderID)
+
+	order.Items = foundOrderItems
+	if err != nil {
+		return nil, err
+	}
+
+	return &order, nil
+}
